Organize folders' structure
